wb-citations-main/requirements.txt
asyncio==3.4.3
Flask==3.0.0
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.12.0
google-cloud-dialogflow-cx==1.28.0
google-cloud-discoveryengine==0.11.2
httpx==0.25.0
langchain==0.0.350
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.0
gunicorn==20.1.0
eventlet==0.30.2
Werkzeug>=3.0.0